ReplaceAliasingEvalWithProject
