+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/optimizer/LogicalPlanOptimizer.java
@@ -35,7 +35,6 @@ import org.elasticsearch.xpack.esql.core.plan.logical.OrderBy;
 import org.elasticsearch.xpack.esql.core.plan.logical.UnaryPlan;
 import org.elasticsearch.xpack.esql.core.rule.ParameterizedRule;
 import org.elasticsearch.xpack.esql.core.rule.ParameterizedRuleExecutor;
-import org.elasticsearch.xpack.esql.core.rule.Rule;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.core.util.CollectionUtils;
@@ -76,6 +75,7 @@ import org.elasticsearch.xpack.esql.optimizer.rules.PushDownEnrich;
 import org.elasticsearch.xpack.esql.optimizer.rules.PushDownEval;
 import org.elasticsearch.xpack.esql.optimizer.rules.PushDownRegexExtract;
 import org.elasticsearch.xpack.esql.optimizer.rules.RemoveStatsOverride;
+import org.elasticsearch.xpack.esql.optimizer.rules.ReplaceAliasingEvalWithProject;
 import org.elasticsearch.xpack.esql.optimizer.rules.SetAsOptimized;
 import org.elasticsearch.xpack.esql.optimizer.rules.SimplifyComparisonsArithmetics;
 import org.elasticsearch.xpack.esql.plan.logical.Aggregate;
@@ -818,86 +818,6 @@ public class LogicalPlanOptimizer extends ParameterizedRuleExecutor<LogicalPlan,
 
         static String syntheticName(Expression expression, Expression af, int counter) {
             return SubstituteSurrogates.temporaryName(expression, af, counter);
-        }
-    }
-
-    /**
-     * Replace aliasing evals (eval x=a) with a projection which can be further combined / simplified.
-     * The rule gets applied only if there's another project (Project/Stats) above it.
-     *
-     * Needs to take into account shadowing of potentially intermediate fields:
-     * eval x = a + 1, y = x, z = y + 1, y = z, w = y + 1
-     * The output should be
-     * eval x = a + 1, z = a + 1 + 1, w = a + 1 + 1
-     * project x, z, z as y, w
-     */
-    static class ReplaceAliasingEvalWithProject extends Rule<LogicalPlan, LogicalPlan> {
-
-        @Override
-        public LogicalPlan apply(LogicalPlan logicalPlan) {
-            Holder<Boolean> enabled = new Holder<>(false);
-
-            return logicalPlan.transformDown(p -> {
-                // found projection, turn enable flag on
-                if (p instanceof Aggregate || p instanceof Project) {
-                    enabled.set(true);
-                } else if (enabled.get() && p instanceof Eval eval) {
-                    p = rule(eval);
-                }
-
-                return p;
-            });
-        }
-
-        private LogicalPlan rule(Eval eval) {
-            LogicalPlan plan = eval;
-
-            // holds simple aliases such as b = a, c = b, d = c
-            AttributeMap<Expression> basicAliases = new AttributeMap<>();
-            // same as above but keeps the original expression
-            AttributeMap<NamedExpression> basicAliasSources = new AttributeMap<>();
-
-            List<Alias> keptFields = new ArrayList<>();
-
-            var fields = eval.fields();
-            for (int i = 0, size = fields.size(); i < size; i++) {
-                Alias field = fields.get(i);
-                Expression child = field.child();
-                var attribute = field.toAttribute();
-                // put the aliases in a separate map to separate the underlying resolve from other aliases
-                if (child instanceof Attribute) {
-                    basicAliases.put(attribute, child);
-                    basicAliasSources.put(attribute, field);
-                } else {
-                    // be lazy and start replacing name aliases only if needed
-                    if (basicAliases.size() > 0) {
-                        // update the child through the field
-                        field = (Alias) field.transformUp(e -> basicAliases.resolve(e, e));
-                    }
-                    keptFields.add(field);
-                }
-            }
-
-            // at least one alias encountered, move it into a project
-            if (basicAliases.size() > 0) {
-                // preserve the eval output (takes care of shadowing and order) but replace the basic aliases
-                List<NamedExpression> projections = new ArrayList<>(eval.output());
-                // replace the removed aliases with their initial definition - however use the output to preserve the shadowing
-                for (int i = projections.size() - 1; i >= 0; i--) {
-                    NamedExpression project = projections.get(i);
-                    projections.set(i, basicAliasSources.getOrDefault(project, project));
-                }
-
-                LogicalPlan child = eval.child();
-                if (keptFields.size() > 0) {
-                    // replace the eval with just the kept fields
-                    child = new Eval(eval.source(), eval.child(), keptFields);
-                }
-                // put the projection in place
-                plan = new Project(eval.source(), child, projections);
-            }
-
-            return plan;
         }
     }
